record objects and compute a database of them for flann matching
the recorder.py can be used to draw a frame around an object and to record
N images from different angles when the object is on a revolving table.
For each object a file with the images and depth information is created. 

create_db.py reads thes object files and computes sift features on them