fix the histogram clustering
there was a very old bug in the histogram clustering. I used the indece of the histogram times 10 as the distance.
But this was only valid in the beginning when the histogram went from 0 to 6000 in 600 bins.

I also added a stronger convolution to the second part of the histogram